Simplify tables and various changes.

+++ src/main/java/com/quiteharmless/mbrweb/bo/Visitor.java
@@ -4,7 +4,7 @@ public class Visitor {
 
 	private String lookupId;
 
-	private int memberId;
+	private Long memberId;
 
 	private String firstName;
 
@@ -16,100 +16,58 @@ public class Visitor {
 
 	private String notes;
 
-	/**
-	 * @return the lookupId
-	 */
 	public String getLookupId() {
 		return lookupId;
 	}
 
-	/**
-	 * @param lookupId the lookupId to set
-	 */
 	public void setLookupId(String lookupId) {
 		this.lookupId = lookupId;
 	}
 
-	/**
-	 * @return the memberId
-	 */
-	public int getMemberId() {
+	public Long getMemberId() {
 		return memberId;
 	}
 
-	/**
-	 * @param memberId the memberId to set
-	 */
-	public void setMemberId(int memberId) {
+	public void setMemberId(Long memberId) {
 		this.memberId = memberId;
 	}
 
-	/**
-	 * @return the firstName
-	 */
 	public String getFirstName() {
 		return firstName;
 	}
 
-	/**
-	 * @param firstName the firstName to set
-	 */
 	public void setFirstName(String firstName) {
 		this.firstName = firstName;
 	}
 
-	/**
-	 * @return the lastName
-	 */
 	public String getLastName() {
 		return lastName;
 	}
 
-	/**
-	 * @param lastName the lastName to set
-	 */
 	public void setLastName(String lastName) {
 		this.lastName = lastName;
 	}
 
-	/**
-	 * @return the timestamp
-	 */
 	public long getTimestamp() {
 		return timestamp;
 	}
 
-	/**
-	 * @param timestamp the timestamp to set
-	 */
 	public void setTimestamp(long timestamp) {
 		this.timestamp = timestamp;
 	}
 
-	/**
-	 * @return the ok
-	 */
 	public boolean isOk() {
 		return ok;
 	}
 
-	/**
-	 * @param ok the ok to set
-	 */
 	public void setOk(boolean ok) {
 		this.ok = ok;
 	}
 
-	/**
-	 * @return the notes
-	 */
 	public String getNotes() {
 		return notes;
 	}
 
-	/**
-	 * @param notes the notes to set
-	 */
 	public void setNotes(String notes) {
 		this.notes = notes;
 	}
