+++ src/main/java/com/quiteharmless/mbrweb/util/Constants.java
@@ -3,7 +3,7 @@ package com.quiteharmless.mbrweb.util;
 public class Constants {
 
 	public static final String UNKNOWN_VISITOR_NAME = "UNKNOWN";
-	public static final int UNKNOWN_VISITOR_ID = 0;
+	public static final Long UNKNOWN_VISITOR_ID = 0L;
 
 	private Constants() {
 	}
